modifies field

+++ api-venv/api/albums/models/track.py
@@ -18,7 +18,7 @@ class Track(models.Model):
     duration = models.IntegerField(default=0, blank=True, null=True)
     tid = models.CharField(default='', unique=True)
     slug = models.SlugField(blank=True, default='')
-    order_no = models.SmallIntegerField(default=0)
+    order_no = models.IntegerField(default=0)
     is_exclusive_sold = models.BooleanField(blank=True, default=False)
     created = models.DateTimeField(auto_now_add=True, null=True)
     updated = models.DateTimeField(auto_now=True, null=True)
